backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
asyncpg==0.30.0
openai==1.109.1
python-jose[cryptography]==3.3.0
python-dotenv==1.1.1
slowapi==0.1.9
pydantic==2.11.7
pytest==8.4.1
pytest-asyncio==1.1.0
